Feat/tests

+++ src/main/java/com/time3/api/domains/ProductOrder/dtos/ProductOrderDto.java
@@ -2,7 +2,8 @@ package com.time3.api.domains.ProductOrder.dtos;
 
 import com.time3.api.domains.Products.SizeEnum;
 
+import java.math.BigInteger;
 import java.util.UUID;
 
-public record ProductOrderDto (UUID productId, Integer quantity, SizeEnum size){
+public record ProductOrderDto (UUID productId, BigInteger quantity, SizeEnum size){
 }
